fix seeders
fix menu manager

add RendersWidgets trait
