+++ database/seeders/BankConnectionTableSeeder.php
@@ -9,6 +9,6 @@ class BankConnectionTableSeeder extends Seeder
 {
     public function run(): void
     {
-        BankConnection::factory()->count(rand(0, 3))->create();
+        BankConnection::factory()->count(rand(1, 3))->create();
     }
 }
