+++ database/seeders/LanguageTableSeeder.php
@@ -9,6 +9,16 @@ class LanguageTableSeeder extends Seeder
 {
     public function run(): void
     {
+        Language::factory()->create([
+            'name' => 'English',
+            'language_code' => 'en',
+            'is_default' => true,
+        ]);
+        Language::factory()->create([
+            'name' => 'Deutsch',
+            'language_code' => 'de',
+        ]);
+
         Language::factory()->count(5)->create();
     }
 }
