+++ database/seeders/UserTableSeeder.php
@@ -6,6 +6,7 @@ use FluxErp\Facades\Widget;
 use FluxErp\Models\Language;
 use FluxErp\Models\User;
 use Illuminate\Database\Seeder;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Hash;
 
 class UserTableSeeder extends Seeder
@@ -22,12 +23,23 @@ class UserTableSeeder extends Seeder
             ])
             ->each(function (User $user) {
                 $user->assignRole('Super Admin');
-                foreach (Widget::all() as $widget) {
+                $row = 0;
+                $col = 0;
+                foreach (Arr::sort(Widget::all(), 'defaultWidth') as $widget) {
                     $user->widgets()->create([
                         'name' => $widget['label'],
                         'component_name' => $widget['component_name'],
-                        'width' => rand(3, 6),
+                        'width' => data_get($widget, 'defaultWidth', 1),
+                        'height' => data_get($widget, 'defaultHeight', 1),
+                        'order_column' => $col,
+                        'order_row' => $row,
                     ]);
+
+                    $col += data_get($widget, 'defaultWidth', 1);
+                    if ($col >= 6) {
+                        $col = 0;
+                        $row++;
+                    }
                 }
             });
     }
